+++ back-end/chatter-be/model/config.js
@@ -1,0 +1,11 @@
+require('dotenv').config();
+
+const dbConfig = {
+  host: process.env.DB_HOST,
+  user: process.env.DB_USER,
+  password: process.env.DB_PASSWORD,
+  port: process.env.DB_PORT,
+  database: process.env.DB_DATABASE
+}
+
+module.exports = dbConfig
