+++ back-end/chatter-be/package.json
@@ -12,6 +12,8 @@
   "dependencies": {
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
+    "express-mysql-session": "^2.1.5",
+    "express-session": "^1.17.1",
     "mysql2": "^2.2.5"
   },
   "devDependencies": {
